% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin.R
\name{bowlingpin}
\alias{bowlingpin}
\title{Main function to run the game for two players}
\usage{
bowlingpin()
}
\value{
A numeric vector containing the total score for each user.
}
\description{
Main function to run the game for two players
}
\examples{
bowlingpin()
}
