% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlingpin.R
\name{bowlingpin}
\alias{bowlingpin}
\title{Main function to run the game for two players}
\usage{
bowlingpin(player1_rolls, player2_rolls)
}
\arguments{
\item{player1_rolls}{A numeric vector having scores for every throw of player 1}

\item{player2_rolls}{A numeric vector having scores for every throw of player 2}
}
\value{
A numeric vector containing the total score for each user.
}
\description{
Main function to run the game for two players
}
\examples{
player1_rolls <- c(2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2)
player2_rolls <- c(2,2,2,2,8,2,6,2,2,2,2,2,5,2,2,2,4,2,2,3)
bowlingpin(player1_rolls,player2_rolls)
}
