% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollinput.R
\name{rollinput}
\alias{rollinput}
\title{Function to take user input for throws for a single player}
\usage{
rollinput(player_num)
}
\arguments{
\item{player_num}{A number indicating if it is player 1 or player 2}
}
\value{
A numeric vector representing the sequence of pins knocked down by the player in each throw.
}
\description{
Function to take user input for throws for a single player
}
\examples{
rolls <- function(){
  player1 <- rollinput(1)
  player2 <- rollinput(2)
}


}
